requirements.txt
fastapi==0.88.0
pymongo==4.1.1
rcon<=2.2.1
steam==1.3.0
uvicorn==0.20.0
discord.py==2.0.0
requests==2.26.0
beautifulsoup4==4.11.1
dnspython==2.2.1
aiohttp==3.8.3
